requirements-dev.txt
django-debug-toolbar>=3.8.1
django-extensions>=3.2.3

# pinned for drf-spectacular
# attrs>=23.1.0,<24.0.0
# rpds-py>=0.10.2
# jsonschema-specifications==2023.07.1
drf-spectacular>=0.27.0,<0.28.0

gitlint>=0.19.1
ipython>=8.19.0,<9.0.0
pre-commit>=3.6.0,<4.0.0